Reimplementing clear
