+++ DeMark/src/main/java/components/DemarkProjectComponent.java
@@ -5,7 +5,11 @@ import com.intellij.openapi.project.Project;
 import com.intellij.openapi.startup.StartupManager;
 import main.java.utils.HighlightUtil;
 import org.jetbrains.annotations.NotNull;
+
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.Stack;
 
 /**
  */
@@ -14,7 +18,7 @@ public class DemarkProjectComponent implements ProjectComponent {
     private final Project project;
     @NotNull
     private final ApplicationComponent applicationComponent;
-    private String message;
+    private Stack<HashMap<Integer, String>> unclearHistory;
 
     /**
      * @param project The current project, i.e. the project which was just opened.
@@ -22,7 +26,7 @@ public class DemarkProjectComponent implements ProjectComponent {
     public DemarkProjectComponent(Project project, @NotNull components.DemarkApplicationComponent applicationComponent) {
         this.project = project;
         this.applicationComponent = applicationComponent;
-        this.message = "Hello there";
+        this.unclearHistory = new Stack<>();
     }
 
     public void initComponent() {
@@ -56,5 +60,22 @@ public class DemarkProjectComponent implements ProjectComponent {
      */
     public String getComponentName() {
         return "DemarkProjectComponent";
+    }
+
+
+    /**
+     * Push the history onto the stack
+     * @param history The cleared lines that were removed
+     */
+    public void pushHistory(HashMap<Integer, String> history) {
+        this.unclearHistory.push(history);
+    }
+
+    /**
+     * Undo the last history clear
+     * @return The last cleared lines
+     */
+    public HashMap<Integer, String> popHistory() {
+        return this.unclearHistory.isEmpty() ? new HashMap<>() : this.unclearHistory.pop();
     }
 }
